+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### v1.5.4
 * Fix crash with Treasure Map and Replay the Spire's portal rooms
+* Fix: Mirror image makes construct render incorrectly (#4)
 
 ### v1.5.3
 * Fix: Ornate Mirror now correctly applies powers to calculate damage.